DocVerse/requirements.txt
# Core packages
paddlepaddle==3.1.0
paddleocr==3.2.0
paddlex[ie,multimodal,ocr,trans]==3.2.0

# OpenAI (required by paddlex[ie])
openai>=1.63

# Database
pymongo==4.5.0

# Environment Config
python-dotenv==1.0.0

# Imaging and OCR
pillow==10.0.0
pdf2image==1.16.3
pytesseract==0.3.10

# API & Backend
fastapi==0.104.0
uvicorn==0.24.0
python-multipart==0.0.6

# Frontend & UI
streamlit==1.28.0

# Utilities
requests==2.31.0
jsonschema==4.20.0
pandas==2.0.3
numpy>=1.19.3,<2.0.0

# Optional dependencies commonly needed by PaddleOCR/PaddleX features
attrdict>=2.0.1
cython>=0.29.32
fire>=0.4.0
fonttools>=4.24.0
imgaug>=0.4.0
lmdb>=1.3.0
pdf2docx>=0.5.6
python-docx>=0.8.11
visualdl>=2.2.0
opencv-python            # Let opencv-contrib-python install the compatible version
opencv-contrib-python    # Remove any version limits for compatibility
